requirements.txt
kaggle==1.5.8
python_telegram_bot==12.8
